pyproject.toml
[project]
name = "superset-spcs-toolkit"
version = "0.1.0"
description = "Superset on Snowpark Container Services helper tooling."
requires-python = ">=3.10"
dependencies = [
    "snowflake-connector-python>=3.6",
    "snowflake-sqlalchemy>=1.5",
    "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[tool.uv]
dev-dependencies = ["pytest>=7.4"]

[tool.pytest.ini_options]
pythonpath = [
    "."
]
